+++ src/DetailsView/components/command-bar-buttons-menu.tsx
@@ -85,7 +85,9 @@ export const CommandBarButtonsMenu = NamedFC<CommandBarButtonsMenuProps>(
                         >
                             <MenuList className={styles.menuList}>
                                 {overflowItems.map((item, index) => (
-                                    <span role="group" key={index}>{item?.children}</span>
+                                    <span role="group" key={index}>
+                                        {item?.children}
+                                    </span>
                                 ))}
                             </MenuList>
                         </MenuPopover>
