+++ src/DetailsView/components/report-export-button.tsx
@@ -17,7 +17,6 @@ export const reportExportButtonAutomationId = 'report-export-button';
 
 export const ReportExportButton = NamedFC<ReportExportButtonProps>('ReportExportButton', props => {
     const exportButtonStyles = useCommandButtonStyle();
-    console.log('props in report export button', props);
     return (
         <InsightsCommandButton
             insightsCommandButtonIconProps={{
